FariseyIOSUI-33_HW2_1 - На проверку
