+++ FariseyIOSUI-33_HW2_1/Navigation/Navigation/Assets.xcassets/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
