+++ FariseyIOSUI-33_HW2_1/Navigation/Navigation/Model.swift
@@ -1,0 +1,21 @@
+//
+//  Model.swift
+//  Navigation
+//
+//  Created by Никита on 21.02.2023.
+//
+
+import Foundation
+
+struct Post {
+    var title: String
+}
+
+func choosen() {
+
+    let choosenPost = Post(title: "Here is choosen post")
+
+    let postViewController = PostViewController()
+
+    postViewController.postTitle = choosenPost
+}
