+++ FariseyIOSUI-33_HW2_1/Navigation/Navigation/SceneDelegate.swift
@@ -1,0 +1,90 @@
+//
+//  SceneDelegate.swift
+//  Navigation
+//
+//  Created by Никита on 12.02.2023.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        
+  /*    let feedViewController = FeedViewController()
+        let profileViewController = ProfileViewController()
+        let postViewController = PostViewController()
+        
+        // UINavigationController
+        
+        let feedNavigation = UINavigationController(rootViewController: feedViewController)
+        feedNavigation.tabBarItem = UITabBarItem(title: "Feed", image: UIImage(systemName: "Feed"), selectedImage: UIImage(systemName: "newspaper.circle"))
+        
+        
+        let profileNavigation = UINavigationController(rootViewController: profileViewController)
+        profileNavigation.tabBarItem = UITabBarItem(title: "Profile", image: UIImage(systemName: "Profile"), selectedImage: UIImage(systemName: "person.crop.circle"))
+        
+        
+        let postNavigation = UINavigationController(rootViewController: profileViewController)
+        postNavigation.tabBarItem = UITabBarItem(title: "Post", image: UIImage(systemName: "Post"), selectedImage: UIImage(systemName: "pencil.circle"))
+        
+        
+        // UITabBarController
+        let tabBarController = UITabBarController()
+        tabBarController.viewControllers = [feedNavigation, profileNavigation, postNavigation]
+        tabBarController.tabBar.backgroundColor = UIColor.green
+   */
+        // UIWindow
+        
+        self.window = UIWindow(windowScene: windowScene)
+        self.window?.rootViewController = TabBarController()
+        self.window?.makeKeyAndVisible()
+        
+        /* let window = UIWindow(windowScene: scene)
+        
+        self.window.rootViewController = TabBarController()
+        
+        self.window = window
+        self.window.makeKeyAndVisible()
+         */
+    }
+    
+    
+    
+    
+    
+//////////
+    func sceneDidDisconnect(_ scene: UIScene) {
+        // Called as the scene is being released by the system.
+        // This occurs shortly after the scene enters the background, or when its session is discarded.
+        // Release any resources associated with this scene that can be re-created the next time the scene connects.
+        // The scene may re-connect later, as its session was not necessarily discarded (see `application:didDiscardSceneSessions` instead).
+    }
+
+    func sceneDidBecomeActive(_ scene: UIScene) {
+        // Called when the scene has moved from an inactive state to an active state.
+        // Use this method to restart any tasks that were paused (or not yet started) when the scene was inactive.
+    }
+
+    func sceneWillResignActive(_ scene: UIScene) {
+        // Called when the scene will move from an active state to an inactive state.
+        // This may occur due to temporary interruptions (ex. an incoming phone call).
+    }
+
+    func sceneWillEnterForeground(_ scene: UIScene) {
+        // Called as the scene transitions from the background to the foreground.
+        // Use this method to undo the changes made on entering the background.
+    }
+
+    func sceneDidEnterBackground(_ scene: UIScene) {
+        // Called as the scene transitions from the foreground to the background.
+        // Use this method to save data, release shared resources, and store enough scene-specific state information
+        // to restore the scene back to its current state.
+    }
+
+
+}
+
