+++ FariseyIOSUI-33_HW2_1/Navigation/Navigation/ViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ViewController.swift
+//  Navigation
+//
+//  Created by Никита on 12.02.2023.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
