--- Navigation/Navigation/ProfileViewController.swift
@@ -1,15 +1,0 @@
-//
-//  ProfileViewController.swift
-//  Navigation
-//
-//  Created by Никита on 17.02.2023.
-//
-import UIKit
-
-class ProfileViewController: UIViewController {
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-    }
-}
