--- Navigation/Navigation/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  Navigation
-//
-//  Created by Никита on 12.02.2023.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
